fix(banking): add lazy imports and code-splitting
