+++ banking/package.json
@@ -41,7 +41,6 @@
     "react-markdown": "^10.1.0",
     "react-router": "^7.15.0",
     "react-router-dom": "^7.15.0",
-    "react-virtuoso": "^4.18.6",
     "rehype-raw": "^7.0.0",
     "remark-gfm": "^4.0.1",
     "sonner": "^2.0.7",
