+++ banking/vite.config.ts
@@ -21,5 +21,35 @@ export default defineConfig({
 		outDir: '../erpnext/public/banking',
 		emptyOutDir: true,
 		target: 'es2015',
+		rollupOptions: {
+			output: {
+				manualChunks(id) {
+					if (!id.includes('node_modules')) {
+						return
+					}
+					if (id.includes('react-dom') || id.includes('/react/')) {
+						return 'vendor-react'
+					}
+					if (id.includes('frappe-react-sdk')) {
+						return 'vendor-frappe'
+					}
+					if (id.includes('@tanstack')) {
+						return 'vendor-tanstack'
+					}
+					if (id.includes('fuse.js')) {
+						return 'vendor-fuse'
+					}
+					if (id.includes('radix-ui') || id.includes('@radix-ui')) {
+						return 'vendor-radix'
+					}
+					if (id.includes('jotai')) {
+						return 'vendor-jotai'
+					}
+					if (id.includes('lucide-react')) {
+						return 'vendor-lucide'
+					}
+				},
+			},
+		},
 	},
 });
